tiles/54/M/TA
Papua New Guinea (PNG; /ˈpæpuə njuː ˈɡɪniː, ˈpɑː-, -pju-/, US: /ˈpæpjuə, pɑːˈpuːə/;[9] Tok Pisin: Papua Niugini; Hiri Motu: Papua Niu Gini), officially the Independent State of Papua New Guinea, is an Oceanian country that occupies the eastern half of the island of New Guinea and its offshore islands in Melanesia, a region of the southwestern Pacific Ocean north of Australia. Its capital, located along its southeastern coast, is Port Moresby. The western half of New Guinea forms the Indonesian provinces of Papua and West Papua.

At the national level, after being ruled by three external powers since 1884, Papua New Guinea established its sovereignty in 1975. This followed nearly 60 years of Australian administration, which started during World War I. It became an independent Commonwealth realm in 1975 with Queen Elizabeth II as its head of state and became a member of the Commonwealth of Nations in its own right.

Papua New Guinea is one of the most culturally diverse countries in the world. There are 852 known languages in the country, of which 12 now have no known living speakers.[10] Most of the population of more than 7 million people live in customary communities, which are as diverse as the languages.[11] It is also one of the most rural, as only 18 percent of its people live in urban centres.[12] The country is one of the world's least explored, culturally and geographically. It is known to have numerous groups of uncontacted peoples, and researchers believe there are many undiscovered species of plants and animals in the interior.[13]

Papua New Guinea is classified as a developing economy by the International Monetary Fund.[14] Strong growth in Papua New Guinea's mining and resource sector led to the country becoming the sixth fastest-growing economy in the world in 2011.[15] Growth was expected to slow once major resource projects came on line in 2015.[16] Mining remains a major economic factor, however. Local and national governments are discussing the potential of resuming mining operations in Panguna mine in Bougainville Province, which has been closed since the civil war in the 1980s–1990s.[17] Nearly 40 percent of the population lives a self-sustainable natural lifestyle with no access to global capital.[18]

Most of the people still live in strong traditional social groups based on farming. Their social lives combine traditional religion with modern practices, including primary education.[11] These societies and clans are explicitly acknowledged by the Papua New Guinea Constitution, which expresses the wish for "traditional villages and communities to remain as viable units of Papua New Guinean society"[19] and protects their continuing importance to local and national community life.